pyproject.toml
[tool.poetry]
name = "sparv-kb-ner-plugin"
version = "0.1.0"
description = ""
authors = ["Kristoffer Andersson <kristoffer.andersson@gu.se>"]
readme = "README.md"
packages = [
    {include = "development"},
    {include = "sbx_ner/kb_ner_plugin", from = "bases"},
    {include = "sparv_kb_ner"},
]

[tool.poetry.dependencies]
python = "^3.10"
sparv-pipeline = {git = "https://github.com/spraakbanken/sparv-pipeline.git", rev = "support-py310"}
transformers = "^4.26.0"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"

[tool.poetry.plugins."sparv.plugin"]
sparv_kb_ner = "sparv_kb_ner"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
